[edit] duplicate code
